When the user refreshes the #dialog-page there are no errors produced.
TODO: refactor the Forbidden_Slot_Array to fetch the forbidden_slots when the #dialog-page is refreshed.
Right now, when the #dialog-page is refreshed, nothing happens basically. There are no error, but the code
doesnt attemps to fetch the corresponding forbidden_slots from the server. This has to change by refactoring the code
that fetches and displays the forbidden_slot data are reusing it when the #dialog-page is refreshed (now its used only when
one of the streets sides is clicked).

To know what forbidden_slots to fetch, you need to add the street-side and block-id to the url like so:
    http://www.parkingplan.ca#dialog-page?street_side=2&street-block=22432 lets say.
This way the code executing after the refresh will know what to get from the server by cheking the url being refreshed.

By the way, dialog-page is the page when the forbidden-slot is displayed (the scheudle, days, dates..)